refactor: extract transformFeatures from useApi composable
- Move transformFeatures() to src/utils/feature-transform.ts
- Remove @turf/area and @turf/boolean-equal imports from useApi.ts
- Pure function is now independently testable

Closes #30
